+++ src/libexec/package-tools/mac/generate.sh
@@ -22,6 +22,7 @@ class Azk < Formula
   def install
     prefix.install Dir['*']
     prefix.install Dir['.nvmrc']
+    prefix.install Dir['.dependencies']
   end
 end